+++ src/domain/journey/space/spaceCalloutPage/SpaceCalloutPage.tsx
@@ -1,11 +1,11 @@
 import CalloutPage from '@/domain/collaboration/CalloutPage/CalloutPage';
-import SpaceSubspacesPage from '../../../space/layout/TabbedSpaceL0/Tabs/SpaceSubspaces/SpaceSubspacesPage';
-import SpaceKnowledgeBasePage from '@/domain/space/layout/TabbedSpaceL0/Tabs/SpaceKnowledgeBase/SpaceKnowledgeBasePage';
+import SpaceSubspacesPage from '../../../space/layout/tabbedLayout/Tabs/SpaceSubspacesPage';
+import SpaceKnowledgeBasePage from '@/domain/space/layout/tabbedLayout/Tabs/SpaceKnowledgeBase/SpaceKnowledgeBasePage';
 import { EntityPageSection } from '@/domain/shared/layout/EntityPageSection';
 import { JourneyCalloutDialogProps } from '@/domain/journey/common/JourneyCalloutDialog/JourneyCalloutDialog';
-import SpaceCommunityPage from '../../../space/layout/TabbedSpaceL0/Tabs/SpaceCommunityPage/SpaceCommunityPage';
+import SpaceCommunityPage from '../../../space/layout/tabbedLayout/Tabs/SpaceCommunityPage/SpaceCommunityPage';
 import { useSpace } from '../../../space/SpaceContext/useSpace';
-import SpaceDashboardPage from '@/domain/space/layout/TabbedSpaceL0/Tabs/SpaceDashboard/SpaceDashboardPage';
+import SpaceDashboardPage from '@/domain/space/layout/tabbedLayout/Tabs/SpaceDashboard/SpaceDashboardPage';
 
 const getPageSection = (position: number | undefined): EntityPageSection => {
   switch (position) {
