+++ src/domain/journey/subspace/routing/SubspaceRoute.tsx
@@ -8,7 +8,7 @@ import { NotFoundPageLayout } from '@/domain/journey/common/EntityPageLayout';
 import { routes } from './challengeRoutes';
 import CalloutRoute from '@/domain/collaboration/callout/routing/CalloutRoute';
 import SubspaceAboutPage from '../../../space/about/SubspaceAboutPage';
-import SubspaceHomePage from '../../../space/layout/SubspaceFlow/SubspaceHomePage';
+import SubspaceHomePage from '../../../space/layout/flowLayout/SubspaceHomePage';
 import Redirect from '@/core/routing/Redirect';
 import { StorageConfigContextProvider } from '@/domain/storage/StorageBucket/StorageConfigContext';
 import SubspaceCalloutPage from '../subspaceCalloutPage/SubspaceCalloutPage';
