+++ src/domain/journey/subspace/subspaceCalloutPage/SubspaceCalloutPage.tsx
@@ -1,5 +1,5 @@
 import CalloutPage from '@/domain/collaboration/CalloutPage/CalloutPage';
-import SubspaceHomePage from '../../../space/layout/SubspaceFlow/SubspaceHomePage';
+import SubspaceHomePage from '../../../space/layout/flowLayout/SubspaceHomePage';
 import { JourneyCalloutDialogProps } from '@/domain/journey/common/JourneyCalloutDialog/JourneyCalloutDialog';
 import { useSubSpace } from '../hooks/useSubSpace';
