+++ src/domain/space/layout/tabbedLayout/SpaceTabProviderQueries.graphql
