+++ src/domain/space/layout/tabbedLayout/Tabs/SpaceCommunityPage/SpaceCommunityPageQueries.graphql
