+++ src/domain/space/layout/tabbedLayout/Tabs/SpaceDashboard/SpaceDashboard.graphql
