+++ src/domain/space/layout/tabbedLayout/Tabs/SpaceTabs.graphql
