+++ src/main/routing/TopLevelRoutes.tsx
@@ -56,10 +56,7 @@ const InnovationHubsRoutes = lazyWithGlobalErrorHandler(
 const CreateSpaceDialog = lazyWithGlobalErrorHandler(
   () => import('@/domain/journey/space/createSpace/CreateSpaceDialog')
 );
-const SpaceRoute = lazyWithGlobalErrorHandler(
-  () => import('@/domain/space/routing/SpaceTabbedLayoutRoute'),
-  withUrlResolverParams
-);
+const SpaceRoute = lazyWithGlobalErrorHandler(() => import('@/domain/space/routing/SpaceRoute'), withUrlResolverParams);
 
 export const TopLevelRoutes = () => {
   useRedirectToIdentityDomain();
